% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcor_PB_adjust.R
\name{gcor_PB_adjust}
\alias{gcor_PB_adjust}
\title{Main function for the adjustment of genetic correlation between 2 phenotypes}
\usage{
gcor_PB_adjust(
  path,
  mean_shift1,
  mean_shift2,
  trait_name1 = "trait1",
  trait_name2 = "trait1",
  trait_binary1 = F,
  trait_binary2 = F,
  K1 = 1,
  K2 = 1,
  h2x = 0.125,
  alpha = 0.055,
  lambda = 2,
  shift_sd1 = 1/sqrt(20208),
  shift_sd2 = 1/sqrt(20208)
)
}
\arguments{
\item{path}{working path}

\item{mean_shift1}{mean shift of the phenotype 1 of interest, between the sample of participant (UKBB) and the population, standardized in the sample of participants ((mean_participants-mean_population)/SE_in_participants)}

\item{mean_shift2}{mean shift of the phenotype 2 of interest, between the sample of participant (UKBB) and the population, standardized in the sample of participants ((mean_participants-mean_population)/SE_in_participants)}

\item{trait_name1}{Name of the phenotype 1}

\item{trait_name2}{Name of the phenotype 2}

\item{trait_binary1}{trait type of phenotype 1, T/F, T for binary traits, F for the other trait types}

\item{trait_binary2}{trait type of phenotype 2, T/F, T for binary traits, F for the other trait types}

\item{K1}{disease prevalence for phenotype 1. Need to be specified for 'binary' phenotypes to derive observed-scale heritability}

\item{K2}{disease prevalence for phenotype 2. Need to be specified for 'binary' phenotypes to derive observed-scale heritability}

\item{h2x}{heritability of participation liability score. Default=0.125 in UKBB}

\item{alpha}{participation rate. Default=0.055 in UKBB}

\item{lambda}{recurrence ratio. Default=2 in UKBB}

\item{shift_sd1}{standard deviations of the mean shift of phenotype 1, default=1/sqrt(HSE_sample_size)}

\item{shift_sd2}{standard deviations of the mean shift of phenotype 2, default=1/sqrt(HSE_sample_size)}
}
\description{
heritability adjustment
}
