% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2_PB_adjust.R
\name{h2_PB_adjust}
\alias{h2_PB_adjust}
\title{Main function for heritability adjustment}
\usage{
h2_PB_adjust(
  path,
  mean_shift,
  trait_name = "trait1",
  trait_binary = F,
  K = 1,
  h2x = 0.125,
  alpha = 0.055,
  lambda = 2,
  shift_sd = 1/sqrt(20208)
)
}
\arguments{
\item{path}{working path}

\item{mean_shift}{mean shift of the phenotype of interest, between the sample of participant (UKBB) and the population, standardized in the sample of participants ((mean_participants-mean_population)/SE_in_participants)}

\item{trait_name}{Name of the phenotype}

\item{trait_binary}{T/F, T for binary traits, F for the other trait types}

\item{K}{disease prevalence. Need to be specified for 'binary' phenotypes to derive observed-scale heritability}

\item{h2x}{heritability of participation liability score. Default=0.125 in UKBB}

\item{alpha}{participation rate. Default=0.055 in UKBB}

\item{lambda}{recurrence ratio. Default=2 in UKBB}

\item{shift_sd}{standard deviations of the mean shift, default=1/sqrt(HSE_sample_size)}
}
\description{
heritability adjustment
}
